requirements.txt
# Source ingestion: PDF → Markdown, remote HTML → mirror + Markdown
# Python 3.10+ recommended.

httpx>=0.27.0
pymupdf4llm>=0.0.17
PyYAML>=6.0.1
trafilatura>=1.8.0

# Heavier, often higher-quality PDF layout (optional alternative to pymupdf4llm):
#   pip install marker-pdf
#   pip install docling
